LayoutTests/inspector/console/console-screenshot-expected.txt

CONSOLE MESSAGE: [object HTMLDivElement]
CONSOLE MESSAGE: [object HTMLDivElement]
CONSOLE MESSAGE: Could not capture screenshot
Tests for the console.screenshot API.


== Running test suite: console.screenshot
-- Running test case: console.screenshot.SingleNode
PASS: The added message should be an image.
PASS: The image should not be empty.

-- Running test case: console.screenshot.MultipleNodes
PASS: The added message should be an image.
PASS: The image should not be empty.

-- Running test case: console.screenshot.DetachedNode
PASS: Could not capture screenshot

-- Running test case: console.screenshot.NoArguments
PASS: The added message should be an image.
PASS: The image should not be empty.
PASS: The image width should be greater than 2px.
PASS: The image height should be greater than 2px.
